+++ src/main/java/com/example/easybooking/user/domain/repository/CustomerProfileRepository.java
@@ -4,4 +4,5 @@ import com.example.easybooking.user.domain.CustomerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerProfileRepository extends JpaRepository<CustomerProfile,Long> {
+    void deleteByProviderId(String providerId);
 }
